[FE][Add pet contract] reorganised form folder and components
